Add game supervisor
Start adding the GameSupervisor. Doesn't recover state yet.

Simplified how the registry_tuple works, so it just takes the public_id rather than sometimes generating one. When the public ID gets generated requires more thought, but I think it will probably happen when a request comes in, so an ID can passed to the GameSupervisor and also back to the users.

Book I'm reading says to use Supervisor but Elixir warnings make clear that Supervisor is now for more static processes and the kind of runtime supervision I need should be handled via DynamicSupervisor. Although it works as is, would be good to refactor to use DynamicSupervisor instead.